[Decomp][Player] Cohesion split: extract death/corpse methods to PlayerDeath.cpp
